+++ src/main/java/dev/jasmine/carrybabyanimals/nursery/NurseryHazard.java
@@ -1,0 +1,9 @@
+package dev.jasmine.carrybabyanimals.nursery;
+
+public enum NurseryHazard {
+    LAVA,
+    FIRE,
+    CACTUS_OR_DAMAGE,
+    SUFFOCATION,
+    DANGEROUS_FALL
+}